Initialize repo, codeboxes <= Ch2, Ex Ch2
Initialized the repository and added a readme file with some information
Still missing is some clear goal in the readme file on how to translate
the whole rethinking package into python, who to ask for help in doing
so, etc.

Added code boxes as ipython notebooks for preface and chapter 2. The
ones directly using the rethinking package are of course not available.

Added exercises chapter 2.